Pass navigation down to root screen layer. Replace pair parameter to string resource. Use collectAsStateWithLifecycle replacing collectAsState. Remove NavigateToCalendar events. Add lifecycle compose runtime dependency.
