+++ app/src/main/java/com/tasky/login/LoginClickableText.kt
@@ -1,5 +1,6 @@
 package com.tasky.login
 
+import androidx.annotation.StringRes
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.ClickableText
@@ -18,7 +19,8 @@ import com.tasky.ui.theme.LightGray3
 
 @Composable
 fun LoginClickableText(
-    linkTextId: Pair<Int, Int>,
+    @StringRes registerInfoText: Int,
+    @StringRes registerActionText: Int,
     onClick: (Int) -> Unit
 ) {
     ClickableText(
@@ -29,12 +31,12 @@ fun LoginClickableText(
             .padding(32.dp),
         text = buildAnnotatedString {
             withStyle(style = SpanStyle(color = LightGray3)) {
-                val text = stringResource(id = linkTextId.first)
+                val text = stringResource(id = registerInfoText)
                 append(text)
             }
             append(" ")
             withStyle(style = SpanStyle(color = Blue)) {
-                val text = stringResource(id = linkTextId.second)
+                val text = stringResource(id = registerActionText)
                 append(text)
             }
         }
