+++ app/src/main/java/com/tasky/login/LoginEvent.kt
@@ -1,6 +1,5 @@
 package com.tasky.login
 
 sealed class LoginEvent {
-    data object NavigateToCalendar : LoginEvent()
     data object NavigateToRegisterScreen : LoginEvent()
 }
