+++ app/src/main/java/com/tasky/register/RegisterEvent.kt
@@ -1,6 +1,5 @@
 package com.tasky.register
 
 sealed class RegisterEvent {
-    data object NavigateToCalendar : RegisterEvent()
     data object NavigateUp : RegisterEvent()
 }
